requirements.txt
aiogram==3.4.1
gspread==6.0.2
google-auth==2.27.0
python-dotenv==1.0.0
aiohttp==3.9.1
